Delete redundant dataset collection screen and abstract preview to seperate component
